+++ test/runtests.jl
@@ -32,6 +32,7 @@ for file in [
     "test_cancel.jl",
     "test_closed.jl",
     "test_remark.jl",
+    "test_rounds.jl",
     "test_quality.jl",
 ]
     path = joinpath(@__DIR__, file)